refactor: 파일/변수 네이밍 변경
